requirements.txt
aiohttp==3.6.2
docker==4.2.0
file:./model/activity#egg=openapi_activity_client
file:./model/market#egg=openapi_market_client
file:./model/payment#egg=openapi_payment_client
func_timeout==4.3.5
mitmproxy>=5.1,<6.0
pytest-asyncio==0.12.0
pytest==5.4.2
pyyaml==5.3.1
transitions==0.8.1
typing_extensions==3.7.4.2
